Move posit types to posit.rs
