+++ chain-signatures/node/src/protocol/mod.rs
@@ -4,6 +4,7 @@ pub mod consensus;
 pub mod contract;
 pub mod error;
 pub mod message;
+pub mod posit;
 pub mod presignature;
 pub mod signature;
 pub mod state;
@@ -15,7 +16,6 @@ pub use contract::primitives::ParticipantInfo;
 pub use contract::ProtocolState;
 pub use cryptography::CryptographicError;
 pub use message::{Message, MessageChannel};
-use serde::{Deserialize, Serialize};
 pub use signature::{IndexedSignRequest, SignQueue};
 pub use state::NodeState;
 
@@ -29,12 +29,9 @@ use crate::storage::presignature_storage::PresignatureStorage;
 use crate::storage::secret_storage::SecretNodeStorageBox;
 use crate::storage::triple_storage::TripleStorage;
 
-use cait_sith::protocol::Participant;
 use near_account_id::AccountId;
 use semver::Version;
-use std::collections::{HashMap, HashSet};
 use std::fmt;
-use std::hash::Hash;
 use std::path::Path;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
@@ -269,140 +266,6 @@ impl Chain {
 impl fmt::Display for Chain {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.as_str())
-    }
-}
-
-/// All actions that can be taken when a new posit is introduced for a protocol.
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-pub enum PositAction {
-    Propose(Vec<Participant>),
-    Accept,
-    // TODO: Reject can also have a reason
-    Reject,
-    /// Aborts the protocol. Only the proposer can send this.
-    Abort,
-}
-
-pub struct PositCounter<Store> {
-    participants: HashSet<Participant>,
-    accepts: HashSet<Participant>,
-    store: Option<Store>,
-}
-
-pub struct Posits<Id, Store> {
-    posits: HashMap<Id, PositCounter<Store>>,
-}
-
-impl<T: Hash + Eq + fmt::Debug, Store> Posits<T, Store> {
-    pub fn new() -> Self {
-        Self {
-            posits: HashMap::new(),
-        }
-    }
-
-    pub fn propose(
-        &mut self,
-        me: Participant,
-        id: T,
-        store: Store,
-        participants: &[Participant],
-    ) -> PositAction {
-        let mut accepts = HashSet::new();
-        accepts.insert(me);
-        self.posits.insert(
-            id,
-            PositCounter {
-                participants: participants.iter().copied().collect(),
-                accepts,
-                store: Some(store),
-            },
-        );
-
-        PositAction::Propose(participants.to_vec())
-    }
-
-    // TODO: make the resp of this synchronous when each of the protocol
-    // managers are their own individual tasks such that they can respond
-    // without the need of the main consensus loop.
-    /// Act on the posit action. This will map the action received to a corresponding
-    /// action to be sent back to the proposer. This will return
-    ///     (ShouldStartProtocol, TemporaryStorageItem, ReplyAction).
-    pub fn act(
-        &mut self,
-        id: T,
-        from: Participant,
-        action: &PositAction,
-        active: &[Participant],
-    ) -> (Option<Vec<Participant>>, Option<Store>, Option<PositAction>) {
-        match action {
-            PositAction::Accept => {
-                let Some(counter) = self.posits.get_mut(&id) else {
-                    tracing::warn!(?id, "received an Accept for a protocol we did NOT propose");
-                    return (None, None, None);
-                };
-
-                counter.accepts.insert(from);
-                let should_start = counter.accepts.len() == counter.participants.len();
-                let should_start =
-                    should_start.then(|| counter.participants.iter().copied().collect());
-
-                let store = if should_start.is_some() {
-                    tracing::info!(?id, "received all Accepts, starting protocol");
-                    self.posits.remove(&id).and_then(|counter| counter.store)
-                } else {
-                    None
-                };
-
-                (should_start, store, None)
-            }
-            PositAction::Reject => {
-                // TODO: On the first reject, we should abort the protocol for now.
-                // We should be able to narrow down the list of participants
-                // that are rejecting the protocol up until the threshold amount.
-                let reply = if self.posits.contains_key(&id) {
-                    Some(PositAction::Abort)
-                } else {
-                    tracing::warn!(?id, "received a Reject for a protocol we did NOT propose");
-                    None
-                };
-                (None, None, reply)
-            }
-            // There's no action to be done here for Abort. Abort should be handled one level above.
-            PositAction::Abort => (None, None, None),
-            PositAction::Propose(participants) => {
-                if self.posits.contains_key(&id) {
-                    tracing::warn!(
-                        ?id,
-                        ?participants,
-                        "received a protocol posit for an id that we already proposed"
-                    );
-                    return (None, None, Some(PositAction::Reject));
-                }
-
-                // Check that the participants are all active
-                for p in participants.iter() {
-                    if !active.contains(p) {
-                        tracing::warn!(?id, ?active, ?participants, "rejecting protocol posit");
-                        return (None, None, Some(PositAction::Reject));
-                    }
-                }
-
-                // Automatically join the protocol if we're accepting here.
-                (Some(participants.to_vec()), None, Some(PositAction::Accept))
-            }
-        }
-    }
-
-    pub fn len(&self) -> usize {
-        self.posits.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.posits.is_empty()
-    }
-
-    pub fn remove(&mut self, id: &T) -> Option<PositCounter<Store>> {
-        self.posits.remove(id)
     }
 }
